Add X10 and DECSET1005 mouse parsing
Got a couple more variants to cover and mouse is good to go.

Good point to break this out into its own library, and write a simple
front end which reports based on what settings are chosen.  Good way to
shake out obvious bugs before setting a fuzzer loose on it.

The long tail of info reports can wait, I've covered what I need to for
linecook, mostly, and want to get back to it.

Signed-off-by: Sam Atman <atmanistan@gmail.com>